+++ src/components/UpvoteSvg.tsx
